--- app_code/netlify/functions/foo.js
@@ -1,3 +1,0 @@
-export default async (request, context) => {
-  return new Response("From netlify foo");
-};
