+++ app_code/package.json
@@ -15,7 +15,6 @@
   "description": "",
   "devDependencies": {
     "hint": "^2.0.0",
-    "netlify-cli": "^20.0.4",
     "sass": "^1.81.0",
     "vercel": "^41.6.2",
     "vite": "^6.3.2"
