--- app_code/src/rolloanvil/origins.js
@@ -1,4 +1,0 @@
-export default {
-  development: "https://testapp8dev.anvil.app",
-  production: "https://testapp8.anvil.app",
-};
